RF FreestyleProjectTest, merging 2 classes into 1 (#1181)
